pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watchclaw"
version = "0.1.0"
description = "Runtime security monitor for AI Agents — Zeek for the AI Agent era"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Eric Bao"}]
keywords = ["ai", "agent", "security", "monitoring", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
llm = [
    "anthropic>=0.20.0",
]

[project.scripts]
watchclaw = "watchclaw.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
